deprecate commands::build_apps
