+++ apps/wsb-task/src/commands/mod.rs
@@ -4,7 +4,6 @@ pub use artifacts::package_artifact;
 pub use artifacts::{app_names, artifact_size, artifacts_dir, executable_names};
 
 mod cargo;
-pub use cargo::build_apps;
 pub use cargo::build_pilot;
 
 pub mod strip;
